Prevent hang probing for The Cricket! in MAME on IIc/IIc+
In MAME, for the SSC in Slot 2, the STATUS flag never comes back with
ready-to-send, so the probe would hang. Introduce a timer, same as on
the receiving side, and timeout instead of hanging.

Similar to https://github.com/a2stuff/a2d/commit/14e72b7384cdbedd3ac5ac6eb287927e8952d7c7